move dataset_layout in core
