+++ athena/core/interfaces/__init__.py
@@ -1,4 +1,4 @@
-from athena.core.interfaces.dataset_layout import DatasetLayout
+from athena.core.dataset_layout import DatasetLayout
 from athena.core.interfaces.fluctuations import Fluctuations
 
 __all__ = ["Fluctuations", "DatasetLayout"]
