+++ athena/core/interfaces/fluctuations.py
@@ -8,7 +8,7 @@ import pandas as pd
 from pydantic import BaseModel, ConfigDict, field_validator, model_validator
 
 from athena.core.market_entities import Candle
-from athena.core.interfaces.dataset_layout import DatasetLayout
+from athena.core.dataset_layout import DatasetLayout
 from athena.core.types import Coin, Period
 
 logger = logging.getLogger(__name__)
